M02: quarantine v2 parameters.py to _v2_legacy/
